Added dependency to pod spec and fixed build warnings

+++ TourDeFranceLineChart/TourDeFranceLineChart/TDFPeakAnnotation.m
@@ -55,7 +55,7 @@ static float const FontSize = 14.f;
     [self updateFrames];
 }
 
-- (void)setElevation:(float)elevation   {
+- (void)setElevation:(CGFloat)elevation   {
     _elevation = elevation;
     
     self.elevationLabel.text = [NSString stringWithFormat:@"%.0f m", elevation];
